+++ src/store/recipes/reducer.js
@@ -5,8 +5,10 @@ const initialState = { imageUrl: null, recipes: [] };
 export default (state = initialState, action) => {
   switch (action.type) {
     case FETCH_RECIPES_SUCCESS:
+      console.log("what is the state before adding recipes", state);
       const { imageUrl, recipes } = action.payload;
       return { imageUrl: imageUrl, recipes: recipes };
+
     default:
       return state;
   }
