+++ src/store/rootReducer.js
@@ -1,4 +1,6 @@
 import { combineReducers } from "redux";
 import recipesReducer from "./recipes/reducer";
 
-export default combineReducers({ recipes: recipesReducer });
+export default combineReducers({
+  recipes: recipesReducer,
+});
